Created the basic 3D View Model based on RayCasting

+++ algoBresenham.h
@@ -17,6 +17,6 @@ void detectLow(int x0, int y0, int x1, int y1, int** listOfBlock, SDL_Renderer* 
 void detect(int x0, int y0, int x1, int y1, int** listOfBlock, int angle, SDL_Renderer* renderer, SDL_Point* point);
 void DDA(float x0, float y0, int x1, int y1, int listOfBlock[SIZE][SIZE], int angle, SDL_Renderer* renderer, SDL_Point* point);
 
-int distance(float x0, float y0, float x1, float y1);
+double distance(float x0, float y0, float x1, float y1);
 
 #endif //RAYCAST3DPROJECT_ALGOBRESENHAM_H
